Use HOC for filtering queue job list
